+++ django_gotolong/demattxn/views.py
@@ -3,6 +3,8 @@
 from django_gotolong.demattxn.models import DematTxn
 
 from django.utils import timezone
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
 
 from django.views import View
 from django.views.generic.list import ListView
@@ -38,11 +40,13 @@ class DematTxnYearArchiveView(YearArchiveView):
         context["year_list"] = self.year_list
         context["month_list"] = self.month_list
         txn_amount = ExpressionWrapper(F('dt_quantity') * F('dt_price'), output_field=IntegerField())
-        total_amount = round(DematTxn.objects.all().filter(dt_date__year=self.get_year()).
+        total_amount = round(DematTxn.objects.all().filter(dt_user_id=self.request.user.id).
+                             filter(dt_date__year=self.get_year()).
                              aggregate(txn_amount=Sum(txn_amount))['txn_amount'])
 
         summary_list = (
-            DematTxn.objects.all().filter(dt_date__year=self.get_year()).
+            DematTxn.objects.all().filter(dt_user_id=self.request.user.id).
+                filter(dt_date__year=self.get_year()).
                 annotate(
                 month=ExtractMonth('dt_date')).values('month').annotate(
                 Total=Round(Sum(txn_amount)))).order_by('month')
@@ -63,21 +67,32 @@ class DematTxnListView(ListView):
 
     # if pagination is desired
     # paginate_by = 300
-    queryset = DematTxn.objects.all()
+    # queryset = DematTxn.objects.all()
 
     year_list = DematTxn.objects.dates('dt_date', 'year')
 
     # month_list = DematTxn.objects.dates('dt_date', 'month')
+
+    def get_queryset(self):
+        return DematTxn.objects.all().filter(dt_user_id=self.request.user.id)
+
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(DematTxnListView, self).dispatch(*args, **kwargs)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["year_list"] = self.year_list
         #  context["month_list"] = self.month_list
         txn_amount = ExpressionWrapper(F('dt_quantity') * F('dt_price'), output_field=IntegerField())
-        total_amount = round(DematTxn.objects.all().aggregate(txn_amount=Sum(txn_amount))[
-                                 'txn_amount'])
+        total_amount = (DematTxn.objects.all().filter(dt_user_id=self.request.user.id).
+            aggregate(txn_amount=Sum(txn_amount))[
+            'txn_amount'])
+        if total_amount:
+            total_amount = round(total_amount)
         summary_list = (
-            DematTxn.objects.all().annotate(
+            DematTxn.objects.all().filter(dt_user_id=self.request.user.id).
+                annotate(
                 year=ExtractYear('dt_date')).values('year').annotate(
                 Total=Round(Sum(txn_amount)))).order_by('year')
         context["total_amount"] = total_amount
@@ -93,6 +108,7 @@ class DematTxnGapView(ListView):
     date_now = timezone.now()
     # date_now = datetime.today().strftime('%Y-%m-%d')
     time_diff = ExpressionWrapper(- F('dt_date'), output_field=fields.DurationField())
+
     # time_diff = ExpressionWrapper(F('due_date'')-Now())
     # ,min_dt_date = Min('dt_date')
     # days=ExpressionWrapper(date_now - F('dt_date'), output_field=fields.DurationField())
@@ -101,12 +117,14 @@ class DematTxnGapView(ListView):
     # annotate(str_datetime=Cast('dt_date', CharField())). \
     #         annotate(str_datetime=Cast('dt_date', CharField())). \
 
-    queryset = DematTxn.objects. \
-        values('dt_ticker'). \
-        annotate(max_dt_date=Max('dt_date')). \
-        order_by('max_dt_date'). \
-        annotate(months_gap=Min(RawSQL('TIMESTAMPDIFF(month, dt_date, curdate())', ()))). \
-        order_by('-months_gap')
+    def get_queryset(self):
+        return DematTxn.objects. \
+            filter(dt_user_id=self.request.user.id). \
+            values('dt_ticker'). \
+            annotate(max_dt_date=Max('dt_date')). \
+            order_by('max_dt_date'). \
+            annotate(months_gap=Min(RawSQL('TIMESTAMPDIFF(month, dt_date, curdate())', ()))). \
+            order_by('-months_gap')
 
     # annotate(first_time_diff=RawSQL('TIMESTAMPDIFF(month, dt_date, curdate())', ())). \
     # annotate(time_diff=Max('first_time_diff')).\
@@ -123,10 +141,13 @@ class DematTxnStatView(ListView):
     # if pagination is desired
     # paginate_by = 300
     txn_amount = ExpressionWrapper(F('dt_quantity') * F('dt_price'), output_field=IntegerField())
-    queryset = DematTxn.objects.all(). \
-        annotate(txn_year=ExtractYear('dt_date')). \
-        values('txn_year').annotate(txn_amount=Round(Sum(txn_amount))). \
-        order_by('txn_year')
+
+    def get_queryset(self):
+        return DematTxn.objects.all(). \
+            filter(dt_user_id=self.request.user.id). \
+            annotate(txn_year=ExtractYear('dt_date')). \
+            values('txn_year').annotate(txn_amount=Round(Sum(txn_amount))). \
+            order_by('txn_year')
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -139,10 +160,13 @@ class DematTxnStatBuySellView(ListView):
     # if pagination is desired
     # paginate_by = 300
     txn_amount = ExpressionWrapper(F('dt_quantity') * F('dt_price'), output_field=IntegerField())
-    queryset = DematTxn.objects.all(). \
-        annotate(txn_year=ExtractYear('dt_date')). \
-        values('txn_year', 'dt_action').annotate(txn_amount=Round(Sum(txn_amount))). \
-        order_by('txn_year')
+
+    def get_queryset(self):
+        return DematTxn.objects.all(). \
+            filter(dt_user_id=self.request.user.id). \
+            annotate(txn_year=ExtractYear('dt_date')). \
+            values('txn_year', 'dt_action').annotate(txn_amount=Round(Sum(txn_amount))). \
+            order_by('txn_year')
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -190,23 +214,30 @@ class DematTxnRefreshView(View):
         template = "gfundareco/gfunda_reco_list.html"
 
         # first delete all existing dematsum objects
-        DematTxn.objects.all().delete()
+        DematTxn.objects.all().filter(dt_user_id=self.request.user.id).delete()
 
-        unique_id = 1
+        max_id_instances = DematTxn.objects.aggregate(max_id=Max('dt_id'))
+        try:
+            max_dt_id = max_id_instances[0].max_id
+        except KeyError:
+            max_dt_id = 0
+
+        unique_id = max_dt_id
         for brec in BrokerIcidirTxn.objects.all():
-            print(brec.stock_symbol, brec.isin_code, brec.quantity)
-            print(brec.txn_price, brec.txn_date)
+            unique_id += 1
+            print(brec.bit_stock_symbol, brec.bit_isin_code, brec.bit_quantity)
+            print(brec.bit_txn_price, brec.bit_txn_date)
             _, created = DematTxn.objects.update_or_create(
                 dt_id=unique_id,
+                dt_user_id=brec.bit_user_id,
                 dt_broker='icidir',
-                dt_ticker=brec.stock_symbol,
-                dt_isin=brec.isin_code,
-                dt_quantity=brec.quantity,
-                dt_price=brec.txn_price,
-                dt_amount=(brec.quantity * brec.txn_price),
-                dt_date=brec.txn_date
+                dt_ticker=brec.bit_stock_symbol,
+                dt_isin=brec.bit_isin_code,
+                dt_quantity=brec.bit_quantity,
+                dt_price=brec.bit_txn_price,
+                dt_amount=(brec.bit_quantity * brec.bit_txn_price),
+                dt_date=brec.bit_txn_date
             )
-            unique_id += 1
 
         # breakpoint()
